Revert ValveTappetService.position to int per APK ground-truth typing
0.8.3 floated position from int to float, citing OpenAPI's generic
"number" type for Thermostat-II position. However, the Bosch APK's own
ValveTappetState client model declares this field as Integer, unlike
sibling TemperatureOffsetState/TemperatureLevelState which do use
Double for their numeric fields — i.e. the app's own ground-truth
model never expects a fractional valve position, contradicting the
loose OpenAPI typing that motivated the float change.

Reverts ValveTappetService.position and SHCThermostat.position back to
int(), with a comment documenting the OpenAPI-vs-APK typing conflict
so a future contributor doesn't re-derive the float assumption from
OpenAPI alone without checking the APK. Updates the corresponding
regression tests (previously asserting float/decimal-preservation) to
assert int, and adds a test documenting the APK-model rationale.

No live-device rawscan was available to directly confirm position
never carries a fractional value; this fix follows the APK's own
typed model as the more authoritative ground truth per project
convention (bosch-shc-api-docs OpenAPI is known to under-specify
numeric fields elsewhere).

+++ boschshcpy/models_impl.py
@@ -1543,7 +1543,7 @@ class SHCThermostat(
         )  # type: ignore[assignment]
 
     @property
-    def position(self) -> float:
+    def position(self) -> int:
         return self._valvetappet_service.position
 
     @property
